Register the extension as a doctree transformer
Using the doctree-read callback causes it to be called after all the doctree
transforms. However, some of these transforms modify the rawsource, such as
the SmartQuote transform. We have to use rawsource because it's the only thing
that doesn't have backslashes stripped out (which is done by the parser, not a
transformer; even with default_priority = 0 it is required).